* '4.7' of scm.dev.nokia.troll.no:qt/oslo-staging-1: (92 commits)
  Make QStackTextEngine cheaper to construct
  Add a function to get the transitions available from a state
  Add a new WA_X11DoNotAcceptFocus attribute for top-level widgets (part 2).
  Code cleaning for the QStringList's joinEmptiness() test case
  Fix behavior change QStringList::join() for null
  Add a new WA_X11DoNotAcceptFocus attribute for top-level widgets.
  Cocoa: fix eventdispatcher crash, found by macgui autotest
  Make QWidget::activateWindow() NET window manager aware.
  Upgraded QLocale data to Unicode CLDR 1.8.0
  Improved CLDR parser.
  Fixed reading draft data from CLDR.
  Improved the cldr parser for QLocale.
  Do not use FSEvents-based filesystemwatcher backend on Mac.
  add file missing in commit 74f5e34979b8a08a91aa3c2fa6d252e68eca7817
  Moc: Add support for rvalue references in signals and slots.
  Add support for polyphonic greek
  Fix build failure on WinCE.
  Autotests: if you use X11 libs, you must link to X11 libs explicitly.
  Disable C++0x mode for QtWebKit and QtScript since WebKit will not compile any time soon with C++0x
  Compile Phonon in C++0x mode.
  ...

+++ src/opengl/qglshaderprogram.cpp
@@ -1290,7 +1290,8 @@ void QGLShaderProgram::setAttributeValue(int location, const QColor& value)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {value.redF(), value.greenF(), value.blueF(), value.alphaF()};
+        GLfloat values[4] = {GLfloat(value.redF()), GLfloat(value.greenF()),
+                             GLfloat(value.blueF()), GLfloat(value.alphaF())};
         glVertexAttrib4fv(location, values);
     }
 }
@@ -2025,7 +2026,8 @@ void QGLShaderProgram::setUniformValue(int location, const QColor& color)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {color.redF(), color.greenF(), color.blueF(), color.alphaF()};
+        GLfloat values[4] = {GLfloat(color.redF()), GLfloat(color.greenF()),
+                             GLfloat(color.blueF()), GLfloat(color.alphaF())};
         glUniform4fv(location, 1, values);
     }
 }
@@ -2054,7 +2056,7 @@ void QGLShaderProgram::setUniformValue(int location, const QPoint& point)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {point.x(), point.y()};
+        GLfloat values[4] = {GLfloat(point.x()), GLfloat(point.y())};
         glUniform2fv(location, 1, values);
     }
 }
@@ -2083,7 +2085,7 @@ void QGLShaderProgram::setUniformValue(int location, const QPointF& point)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {point.x(), point.y()};
+        GLfloat values[4] = {GLfloat(point.x()), GLfloat(point.y())};
         glUniform2fv(location, 1, values);
     }
 }
@@ -2112,7 +2114,7 @@ void QGLShaderProgram::setUniformValue(int location, const QSize& size)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {size.width(), size.width()};
+        GLfloat values[4] = {GLfloat(size.width()), GLfloat(size.width())};
         glUniform2fv(location, 1, values);
     }
 }
@@ -2141,7 +2143,7 @@ void QGLShaderProgram::setUniformValue(int location, const QSizeF& size)
     Q_D(QGLShaderProgram);
     Q_UNUSED(d);
     if (location != -1) {
-        GLfloat values[4] = {size.width(), size.height()};
+        GLfloat values[4] = {GLfloat(size.width()), GLfloat(size.height())};
         glUniform2fv(location, 1, values);
     }
 }
@@ -2562,9 +2564,9 @@ void QGLShaderProgram::setUniformValue(int location, const QTransform& value)
     Q_UNUSED(d);
     if (location != -1) {
         GLfloat mat[3][3] = {
-            {value.m11(), value.m12(), value.m13()},
-            {value.m21(), value.m22(), value.m23()},
-            {value.m31(), value.m32(), value.m33()}
+            {GLfloat(value.m11()), GLfloat(value.m12()), GLfloat(value.m13())},
+            {GLfloat(value.m21()), GLfloat(value.m22()), GLfloat(value.m23())},
+            {GLfloat(value.m31()), GLfloat(value.m32()), GLfloat(value.m33())}
         };
         glUniformMatrix3fv(location, 1, GL_FALSE, mat[0]);
     }
